pyproject.toml
[project]
name = "badgegen"
version = "0.1.0"
description = "Generatore flessibile di badge PDF da Excel e template PDF"
readme = "README.md"
authors = [{ name = "Alex Citeroni" }]
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
  "pymupdf>=1.24.0",
  "pandas>=2.0.0",
  "openpyxl>=3.1.0",
  "typer>=0.9.0",
  "pyyaml>=6.0.0",
]

[project.scripts]
badgegen = "badgegen.cli:main"

[tool.pytest.ini_options]
addopts = "-q"

[tool.black]
line-length = 100
